+++ include/Mouse.h
@@ -18,7 +18,7 @@ public:
 	int getScore() const;
 
 	void addSouls(int soulNum);
-	void move(sf::Vector2f direction, float deltaTime);
+	void move(sf::Vector2i direction, float deltaTime);
 	void draw(sf::RenderWindow * window) const override;
 	void handleKeys(float deltaTime);
